+++ mackup/config.py
@@ -12,6 +12,7 @@ from .constants import (
     ENGINE_COPY,
     ENGINE_ICLOUD,
     ENGINE_FS,
+    ENGINE_MSONEDRIVE,
 )
 
 from .utils import (
@@ -20,6 +21,7 @@ from .utils import (
     get_copy_folder_location,
     get_google_drive_folder_location,
     get_icloud_folder_location,
+    get_microsoft_onedrive_folder_location,
 )
 
 try:
@@ -71,7 +73,7 @@ class Config(object):
         """
         The engine used by the storage.
 
-        ENGINE_DROPBOX, ENGINE_GDRIVE, ENGINE_COPY, ENGINE_ICLOUD or ENGINE_FS.
+        ENGINE_DROPBOX, ENGINE_GDRIVE, ENGINE_COPY, ENGINE_ICLOUD, ENGINE_MSONEDRIVE, or ENGINE_FS.
 
         Returns:
             str
@@ -207,6 +209,7 @@ class Config(object):
             ENGINE_COPY,
             ENGINE_ICLOUD,
             ENGINE_FS,
+            ENGINE_MSONEDRIVE,
         ]:
             raise ConfigError("Unknown storage engine: {}".format(engine))
 
@@ -227,6 +230,8 @@ class Config(object):
             path = get_copy_folder_location()
         elif self.engine == ENGINE_ICLOUD:
             path = get_icloud_folder_location()
+        elif self.engine == ENGINE_MSONEDRIVE:
+            path = get_microsoft_onedrive_folder_location()
         elif self.engine == ENGINE_FS:
             if self._parser.has_option("storage", "path"):
                 cfg_path = self._parser.get("storage", "path")